+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * [#285](https://github.com/clojure-emacs/orchard/issues/285): **BREAKING:** Remove special handling of Boot classpath.
+* [#287](https://github.com/clojure-emacs/orchard/issues/287): Inspector: ton't crash when field contains non-equiv()able value.
 
 ## 0.26.3 (2024-08-14)